Вынес LayoutedWord в отдельный файл
